+++ src/protontricks/cli.py
@@ -10,16 +10,12 @@
 
 import sys
 import argparse
-import shutil
-import subprocess
 import os
-import shlex
 import logging
 
 from . import __version__
 from .steam import (find_proton_app, find_steam_path, find_steam_runtime_path,
-                    get_steam_apps, get_steam_lib_paths,
-                    get_custom_proton_installations)
+                    get_steam_apps, get_steam_lib_paths)
 from .winetricks import get_winetricks_path
 from .gui import select_steam_app_with_gui
 from .util import run_command
@@ -157,20 +153,25 @@ def main():
     # 5. Find any Steam apps
     steam_apps = get_steam_apps(steam_root, steam_lib_paths)
 
-    # it's too early to find proton here, as it cannot be found if no globally active proton version is set.
-    # having no proton at this point is no problem as:
-    # 	1. not all commands require proton(search)
-    #	2. a specific steam-app will be chosen in gui mode, which might use a different proton version than the one found here
+    # It's too early to find Proton here,
+    # as it cannot be found if no globally active Proton version is set.
+    # Having no Proton at this point is no problem as:
+    # 1. not all commands require Proton (search)
+    # 2. a specific steam-app will be chosen in GUI mode,
+    #    which might use a different proton version than the one found here
 
     # Run the GUI
     if args.gui:
         steam_app = select_steam_app_with_gui(steam_apps=steam_apps)
-        # 6. Find proton version of selected app
+
+        # 6. Find Proton version of selected app
         proton_app = find_proton_app(
-        	steam_path=steam_path, steam_apps=steam_apps, appid=steam_app.appid)
+            steam_path=steam_path, steam_apps=steam_apps, appid=steam_app.appid
+        )
         if not proton_app:
-        	print("Proton installation could not be found!")
-        	sys.exit(-1)
+            print("Proton installation could not be found!")
+            sys.exit(-1)
+
         run_command(
             steam_path=steam_path,
             winetricks_path=winetricks_path,
@@ -211,13 +212,15 @@ def main():
             "can find the game."
         )
         return
-	# 6. Find globally active Proton version now
+
+    # 6. Find globally active Proton version now
     proton_app = find_proton_app(
         steam_path=steam_path, steam_apps=steam_apps, appid=args.appid)
 
     if not proton_app:
         print("Proton installation could not be found!")
         sys.exit(-1)
+
     # If neither search or GUI are set, do a normal Winetricks command
     # Find game by appid
     steam_appid = int(args.appid)