+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Changed
+ - Delay Proton detection until it's necessary.
+
+### Fixed
+ - Use the correct Proton installation when selecting a Steam app using the GUI.
+
 ## [1.2] - 2019-02-27
 ### Added
  - Add a `-c` parameter to run shell commands in the game's installation directory with relevant Wine environment variables.