+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
  - Use the correct Proton installation when selecting a Steam app using the GUI.
+ - Print a proper error message if Steam isn't found.
 
 ## [1.2] - 2019-02-27
 ### Added