+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
  - Use the correct Proton installation when selecting a Steam app using the GUI.
  - Print a proper error message if Steam isn't found.
  - Print an error message when GUI is enabled and no games were found.
+ - Support appmanifest files with mixed case field names.
 
 ## [1.2] - 2019-02-27
 ### Added