+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.2.1] - 2019-04-08
 ### Changed
  - Delay Proton detection until it's necessary.
 
@@ -54,7 +54,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - The last release of protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
 
-[Unreleased]: https://github.com/Matoking/protontricks/compare/1.2...HEAD
+[Unreleased]: https://github.com/Matoking/protontricks/compare/1.2.1...HEAD
+[1.2.1]: https://github.com/Matoking/protontricks/compare/1.2...1.2.1
 [1.2]: https://github.com/Matoking/protontricks/compare/1.1.1...1.2
 [1.1.1]: https://github.com/Matoking/protontricks/compare/1.1...1.1.1
 [1.1]: https://github.com/Matoking/protontricks/compare/1.0...1.1